Calculate witness path for tree
This was trickier than I expected, but it resulted in some code cleanup,
so that's good.

I got really confused as to how long witness_path should be.
In a 32 level tree, should it be 31 or 32 elements long?
Some unit tests in sapling-crypto use 32 elements,
but the official zcash code appears to use 31.

32 elements doesn't make sense if the tree has 32 levels,
since there is no "sibling" for the root hash, and no way to verify it.

Pretty sure I got this right, but will revisit
if the code complains later.