Add serialization and enforce trait
This is it! The vector Merkle tree now fulfills the MerkleTree trait.
The last bit was to add some serialization methods.

It might have been valuable to depend on the speedy crate here,
but that would really be done
better if you did speedy all the
way down into the pairing library.

This is not an ideal implementation,
but it gets the job done.
Since the vector based merkle tree is inefficient anyway,
getting the job done is sufficient.