Split out MerkleHasher trait
My decision to put the read and combine_hash
methods as associated methods on HashableElement
was naive.
The problem is that hashing an element may require
additional state.
In our crypto work based on sapling, for example,
we need to have parameters that set up the elliptic curve
used for hashing.

Pulling it out like this was tedious, and it makes the API less pretty.
But it will be more flexible for our needs in the future.