Add depth to combine_hash
This was a hugely annoying refactor.
Thank god for unit tests.
Even though I had to edit them all.

I forgot when designing the API that combine_hash in production
incorporates the depth of the tree at each hash
as a sort of hash personalization.
This needed to be added, which ended up meaning a lot
of tests and code needed editing.

But tests pass so.... it works?