Set vector tree depth to 33
So, this is a weird one.
My zk proofs failed using this tree
because the auth path only had 31 elements.

But everywhere in the sapling code, the tree depth is 32.
My conclusion is that sapling sets tree depth to not count the
root hash.
In my mind that makes no sense,
so I set the tree depth to 33 including
the root hash.

Then my example code worked, so I'm assuming this is correct.