Give up onto the into_iter API.
So iteration is really hard in Rust, how odd.
I was under the impression that I should use IntoIter when possible,
but I have had a lot of trouble with the
nested iterators and this
is the least bad solution I've found.
I'm using it in other crates and wanted to be consistent,
although the main driver behind this change
is that IntoIterator was getting irritating
in the beanchain crate.