Add a truncate method
The implementation recalculates all interim hashes, which is not ideal,
but can be improved later.
(None of the vector implementation is designed to be ideal).

This feature is needed for supporting a global tree
in a blockchain.
If a new block is added to the tree and either it doesn't agree
with the hash,
or the block later has to be discarded,
then the chain needs to roll the tree back
to the state before that block was added.