Store tree size as part of witness.
We've determined that it's easiest to check that a root
existed at some point in the past
in the tree, if we know how big
the tree was
at that past time.

Adding this piece of information to the witness is trivial, and makes
the API easier to use upstream.