Add get method to tree
Allows the user to request a specific
leaf node by position.

I wanted to avoid this because position may not
be a well-defined concept depending
on the type of Merkle Tree being wrapped.

However, for our purposes, position is important.
Peers will want to ask, "What is the leaf node at position x"
and miners have to be able to answer.