Split tests out to separate file
I would've done this from the start, if I'd known it was possible.
This was one of my first projects and I
wasn't very comfortable with rust yet.

I'm making this change now in preparation for
a new Michael tree implementation that is a little seen her around memory management