Add past_root implementation
walk up the left tree and rebuild the root
as though there are no nodes for the right.

When the fake root node is discovered hash it together with itself
until the desired depth has been reached.

there is a lot of symmetry between this function and the root_hash
 and rehash functions.
I think the witness function will also have the symmetry.
It would be good to abstract the walking part from the business logic
using a closure.
But I don't feel like working it out just now!