Remove "new" from trait
The next version of the Merkle tree
will need to have a configuration
directory passed in
for storage.

I imagine there will be multiple ways to construct trees in the future.

I believe I originally put new on the trait for a reason,
though I can't remember what it is!
Probably I thought it would make the trait more generic,
because you wouldn't have to construct a specific one
in client code.
This is basically just a misunderstanding of how traits are used,
from my early days as a Rust developer.