Initial stages of a rocksdb-based merkle tree
The advantage here is that the whole tree doesn't
need to be loaded into memory,
and we don't need to manage the filesystem.

The disadvantage is that it's a lightweight wrapper of a
c++ API,
and I am calling unwrap() everywhere.
I don't know how else to do this,
except by making the trait return
results more often.

This is very partial, as I'm shifting to another task for a bit.