Implement add, truncate for rocks db
Just fleshing out the methods
for this new implementation.

This is functionally similar to the linked list implementation,
but because I don't store a list of nodes directly,
there is some massaging of data.

One thing I'm worried about is the hanging around of
nodes that have been deleted
by truncate.
I don't think there will be any problem with that,
as the keys get overwritten
when the list grows afresh.
However, there could be some inconsistency that I miss.