Sled DB based merkle tree
This started as a port from the rocksdb
implementation to sled db.
I don't trust the rocksdb version
because it doesn't have transaction support yet.
I also think the sled API is neater,
being designed for Rust,
rather than just wrapping C++ Rocks.

The port ended up having a huge refactor,
because the transaction support in sled has some
specific API characteristics.
I also cleaned up the db schema a bit,
not that it's that complicated,
and made everything a bit tidier.

Note: Be aware that sleddb transactions are not reentrant.
If you call a function that creates a transaction from inside a transaction,
you will end up deadlocked.